Improved hit rate on Bio's
Turns out `bio` is a small word that even if matched with `bio` will only return a fuzzy result of 11, which is cut off by our algorightm as being too low.
For this reason the bio field was not filled if its just called `bio`.
Added `Speaker Bio` as an alternate name and this makes it get 4K+ results.